requirements.txt
Django>=4.2,<5.0
django-environ
djangorestframework
django-filter
psycopg2-binary
whitenoise
celery
redis
razorpay
django-cors-headers
